fix: enforce root privileges at the script entry point
