+++ version.py
@@ -1,2 +1,2 @@
 # Cloudflare-Utils version file
-__version__ = "2.6.5-dev3"
+__version__ = "2.6.5-dev4"
